requirements.txt
matplotlib==3.2.2
numpy==1.19.5
pandas==1.1.5
pingouin==0.3.12
scikit_learn==1.0.1
scipy==1.4.1
seaborn==0.11.2
sklearn_genetic==0.4.1
statsmodels==0.10.2
xgboost==0.90
notebook==6.4.5
xlrd==1.2.0